Azure-AI-Foundry-LLM-Lab1/requirements.txt
# Azure AI Foundry dependencies
openai>=1.0.0
python-dotenv>=1.0.0
